Absorb sensor into presentation facade
