+++ crates/perfgate-sensor/Cargo.toml
@@ -7,21 +7,20 @@ license.workspace = true
 repository.workspace = true
 homepage.workspace = true
 authors.workspace = true
-description = "Sensor report building for cockpit integration"
+description = "Workspace-only compatibility wrapper for perfgate::presentation::sensor"
 readme = "README.md"
 documentation = "https://docs.rs/perfgate-sensor"
 keywords = ["sensor", "report", "cockpit", "integration"]
 categories = ["development-tools"]
+publish = false
 
 [dependencies]
-perfgate-types.workspace = true
-serde.workspace = true
-serde_json.workspace = true
+perfgate-app.workspace = true
 
 [dev-dependencies]
-proptest = "1.7"
-insta.workspace = true
+perfgate.workspace = true
 perfgate-types.workspace = true
+serde_json.workspace = true
 
 [lints]
 workspace = true
