+++ crates/perfgate/Cargo.toml
@@ -18,7 +18,6 @@ perfgate-types.workspace = true
 perfgate-domain.workspace = true
 perfgate-adapters.workspace = true
 perfgate-app.workspace = true
-perfgate-sensor.workspace = true
 anyhow.workspace = true
 regex.workspace = true
 serde.workspace = true
