+++ fuzz/Cargo.toml
@@ -19,7 +19,6 @@ perfgate-domain = { path = "../crates/perfgate-domain" }
 perfgate-app = { path = "../crates/perfgate-app" }
 perfgate-export = { path = "../crates/perfgate-export" }
 perfgate-paired = { path = "../crates/perfgate-paired" }
-perfgate-sensor = { path = "../crates/perfgate-sensor" }
 
 [[bin]]
 name = "parse_run_receipt"
