+++ policy/absorbed_crates.txt
@@ -31,7 +31,7 @@ perfgate-sha256 -> perfgate_types::fingerprint; perfgate::core::fingerprint faca
 # Presentation.
 perfgate-render -> perfgate::presentation::render [compatibility wrapper] [publish=false]
 perfgate-export -> perfgate::presentation::export [compatibility wrapper] [publish=false]
-perfgate-sensor -> perfgate::presentation::sensor
+perfgate-sensor -> perfgate::presentation::sensor [compatibility wrapper] [publish=false]
 
 # Runtime, app, and integrations.
 perfgate-adapters -> perfgate::runtime
