+++ tests/cucumber.rs
@@ -28,6 +28,7 @@ use perfgate_types::{
 // Microcrate imports for direct testing
 use perfgate::presentation::export::{ExportFormat, ExportUseCase};
 use perfgate::presentation::render::render_markdown;
+use perfgate::presentation::sensor::SensorReportBuilder;
 use perfgate_domain::budget::{
     aggregate_verdict, evaluate_budget, reason_token as budget_reason_token,
 };
@@ -38,7 +39,6 @@ use perfgate_error::{
     AdapterError, ConfigValidationError, IoError, PairedError, PerfgateError, StatsError,
     ValidationError,
 };
-use perfgate_sensor::SensorReportBuilder;
 use perfgate_types::fingerprint::sha256_hex;
 use perfgate_types::validation::validate_bench_name as validate_bench_name_fn;
